refractor test views and update comments
